refactor: isolate OAuth credential compatibility at the storage boundary
